Internet Chat Application with a server and a client program
The server manages a chat group, allowing any number of clients to join
the group with a user name at
any time. A client can broadcast message or file, unicast message or
file or block cast to any of the connected client. The
project was developed using socket programming in JAVA.